Move "BodyCodegen" into a file with the appropriate name.
Catching up with coding conventions that changed 20+ years ago.
